move T2-MI code to t2mi.c
